requirements.txt
flask
flask_sqlalchemy
Flask-Migrate
